refactor(refl)!: define strict ref and value ownership contracts
- add stable type ids, type ops, and aligned value storage

- enforce const-aware Ref access and recoverable callable validation

- share argument binding between methods and constructors

- sync ECS resources, scripting, samples, and tests with the new contracts

BREAKING CHANGE: refl Ref, Val, Type, callable, and property APIs now enforce stricter type and const semantics.

+++ src/scripting/scripting_engine.cpp
@@ -141,7 +141,7 @@ bool ScriptingEngine::call_function(
     for (const auto& arg : args) {
         lua_push_ref(L, arg);
     }
-    if (lua_pcall(L, args.size(), 0, 0) != LUA_OK) {
+    if (lua_pcall(L, static_cast<int>(args.size()), 0, 0) != LUA_OK) {
         error("Lua call failed: {}", lua_tostring(L, -1));
         lua_pop(L, 1); // Pop the error message
         return false;
